auto refresh after adding custom labels
